MAJOR RECONFIGURATION of chart_example.ipynb:
**********       Raw data and/or the statistics can be written out in a csv or xlsx (or none), select between summarized or raw data values, can include or exclude weather data (user follows instructions in code commentary to assign 3 new variables at the top of the code before running)        **********

Primarily automation, any scenario for any length of time will work as long as variables are correctly specified at top of code

Zeros are artificially inserted into the dpu concentration datasets (documented in notes) by extracting the harvest date and inserting the zero after the harvest date, which allows the harvest date to change without impacting the code.

No spatial data are used or created in the charting process

Code now accesses the temp and precip data (user can pick one or the other at the top of script), summarizes weather weekly, and plots it on a secondary y axis.

Two plots are generated, one with and one without the weather data, single field plots and additional configurations coming soon.

Currently, user can specify 1-4 DPU concentration datasets of interest and they will all be plotted on the same line chart. Code structure in progress to create a framework that handles any number of datasets specified, and comments explain that the user needs to make sure the Y-axis units of the selected variables are compatible (using the hdf5 datastore viewer notebook which prints the units).